+++ lib/src/cleanup.dart
@@ -4,12 +4,13 @@ import 'package:tavern/src/extensions.dart';
 PostProcessBuilder cleanupBuilder(_) => CleanupBuilder();
 
 class CleanupBuilder extends FileDeletingBuilder implements PostProcessBuilder {
-  CleanupBuilder() : super([
-    Extensions.htmlContent,
-    Extensions.markdown,
-    Extensions.markdownContent,
-    Extensions.metadata,
-    Extensions.mustache,
-    Extensions.withPartials,
-  ]);
+  CleanupBuilder()
+      : super(const [
+          Extensions.htmlContent,
+          Extensions.markdown,
+          Extensions.markdownContent,
+          Extensions.metadata,
+          Extensions.mustache,
+          Extensions.withPartials,
+        ]);
 }