+++ lib/src/metadata.dart
@@ -10,6 +10,7 @@ import 'package:tavern/src/utils.dart';
 Builder metadataBuilder(_) => MetadataBuilder();
 
 class MetadataBuilder implements Builder {
+  @override
   Future build(BuildStep buildStep) async {
     var inputId = buildStep.inputId;
 
@@ -29,8 +30,9 @@ class MetadataBuilder implements Builder {
     ]);
   }
 
+  @override
   Map<String, List<String>> get buildExtensions {
-    return {
+    return const {
       Extensions.markdown: [
         Extensions.metadata,
         Extensions.markdownContent,
@@ -62,11 +64,11 @@ MetadataOutput extractMetadata(String fileContents, String path) {
     throw ('unexpected metadata');
   }
 
-  var metadata = new Map<String, dynamic>.from(yaml);
+  var metadata = Map<String, dynamic>.from(yaml);
   metadata['url'] = getHtmlPath(path);
 
   lines.removeRange(first, last + 1);
-  return new MetadataOutput(metadata, lines.join('\n'));
+  return MetadataOutput(metadata, lines.join('\n'));
 }
 
 class MetadataOutput {