+++ lib/src/mustache.dart
@@ -3,25 +3,25 @@ import 'dart:convert';
 import 'package:build/build.dart';
 import 'package:glob/glob.dart';
 import 'package:tavern/src/extensions.dart';
-import 'package:mustache/mustache.dart' as mustache;
+import 'package:mustache_template/mustache.dart' as mustache;
 
 Builder mustacheBuilder(_) => MustacheBuilder();
 
 class MustacheBuilder implements Builder {
+  @override
   Future build(BuildStep buildStep) async {
     var inputId = buildStep.inputId;
 
     var outputId = inputId.changeExtension(Extensions.html);
     var contents = await buildStep.readAsString(inputId);
     var metadata = await _readMetadata(buildStep);
-    var templateName = metadata['template'] ?? "";
+    var templateName = metadata['template'] ?? '';
     var templateStr = await _readTemplate(buildStep, templateName);
 
-
-    var template = new mustache.Template(templateStr, lenient: true);
+    var template = mustache.Template(templateStr, lenient: true);
 
     // also render the metadata to the input file
-    var contentTemplate = new mustache.Template(contents, lenient: true);
+    var contentTemplate = mustache.Template(contents, lenient: true);
     metadata['content'] = contentTemplate.renderString(metadata);
 
     var output = template.renderString(metadata);
@@ -30,20 +30,21 @@ class MustacheBuilder implements Builder {
   }
 
   Future<String> _readTemplate(BuildStep buildStep, String fileName) async {
-    var assets = await buildStep.findAssets(Glob("**.mustache")).toList();
+    var assets = await buildStep.findAssets(Glob('**.mustache')).toList();
     for (var asset in assets) {
       var assetFileName = asset.path;
       if (assetFileName == fileName) {
-        var assetStr = await buildStep
-            .readAsString(new AssetId(asset.package, asset.path));
+        var assetStr =
+            await buildStep.readAsString(AssetId(asset.package, asset.path));
         return assetStr;
       }
     }
 
-    return "";
+    return '';
   }
 
-  Map<String, List<String>> get buildExtensions => {
+  @override
+  Map<String, List<String>> get buildExtensions => const {
         Extensions.htmlContent: [Extensions.html],
       };
 
@@ -52,19 +53,15 @@ class MustacheBuilder implements Builder {
     try {
       return _parseNonNull(await buildStep.readAsString(id));
     } on AssetNotFoundException {
-      return <String, dynamic>{};
+      return {};
     }
   }
 
   static Map<String, dynamic> _parseNonNull(String metadata) {
     try {
-      var m = json.decode(metadata);
-      if (m == null) {
-        return <String, dynamic>{};
-      }
-      return m;
+      return json.decode(metadata) ?? {};
     } on FormatException {
-      return <String, dynamic>{};
+      return {};
     }
   }
 }