+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.3.0
+
+- Update SDK restraint to >= 2.6.0
+- Switch from mustache to mustache_template package
+- Update markdown package to 3.0.0
+- Clean up various lints
+
 ## 1.0.0
 
 - Initial version, created by Stagehand